Moved map def
