+++ ngraph/python/src/pyngraph/pyngraph.cpp
@@ -34,6 +34,7 @@
 #include "pyngraph/ops/util/regmodule_pyngraph_op_util.hpp"
 #include "pyngraph/partial_shape.hpp"
 #include "pyngraph/passes/regmodule_pyngraph_passes.hpp"
+#include "pyngraph/rt_map.hpp"
 #include "pyngraph/shape.hpp"
 #include "pyngraph/strides.hpp"
 #include "pyngraph/types/regmodule_pyngraph_types.hpp"
@@ -45,6 +46,7 @@ namespace py = pybind11;
 PYBIND11_MODULE(_pyngraph, m)
 {
     m.doc() = "Package ngraph.impl that wraps nGraph's namespace ngraph";
+    regclass_pyngraph_PyRTMap(m);
     regclass_pyngraph_Node(m);
     regclass_pyngraph_Input(m);
     regclass_pyngraph_Output(m);
